requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Pillow==11.0.0
python-magic==0.4.27
Flask-Limiter==3.8.0
gunicorn==23.0.0
opencv-python==4.10.0.84
